refactor: change file name of f_cache to ss_cache to easier link to softsim
