+++ lib/ss_cache.h
